move cluster_control from core/control to core/job
