+++ oneflow/core/job/cluster.cpp
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 #include "oneflow/core/job/cluster.h"
-#include "oneflow/core/control/cluster_control.pb.h"
-#include "oneflow/core/control/cluster_control.h"
+#include "oneflow/core/job/cluster_control.pb.h"
+#include "oneflow/core/job/cluster_control.h"
 #include "oneflow/core/control/ctrl_client.h"
 #include "oneflow/core/job/oneflow.h"
 #include "oneflow/core/job/machine_context.h"
