+++ oneflow/core/job/cluster_control.h
@@ -13,8 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_CONTROL_CLUSTER_CONTROL_H_
-#define ONEFLOW_CORE_CONTROL_CLUSTER_CONTROL_H_
+#ifndef ONEFLOW_CORE_JOB_CLUSTER_CONTROL_H_
+#define ONEFLOW_CORE_JOB_CLUSTER_CONTROL_H_
 
 namespace oneflow {
 
@@ -27,4 +27,4 @@ struct ClusterControl final {
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_CONTROL_CLUSTER_CONTROL_H_
+#endif  // ONEFLOW_CORE_JOB_CLUSTER_CONTROL_H_
