+++ oneflow/python/oneflow_internal_helper.h
@@ -28,9 +28,9 @@ limitations under the License.
 #include "oneflow/core/job/machine_context.h"
 #include "oneflow/core/job/oneflow.h"
 #include "oneflow/core/job/runtime_job_descs.h"
-#include "oneflow/core/control/cluster_control.pb.h"
+#include "oneflow/core/job/cluster_control.pb.h"
 #include "oneflow/core/control/ctrl_client.h"
-#include "oneflow/core/control/cluster_control.h"
+#include "oneflow/core/job/cluster_control.h"
 #include "oneflow/core/job/job_build_and_infer_ctx_mgr.h"
 #include "oneflow/core/job/foreign_watcher.h"
 #include "oneflow/core/job/foreign_callback.h"
